Use `buildSubgraph` method to parse/build fed2 subgraphs
This commits slightly update the code building fed2 subgraph schema to
mimick what the fed2 uses. In practice, the only visible changes
compared to the previous code is that the subgraph name will now
be used when pointing to location errors. Meaning that instead of:
```
GraphQL Request:10:20
9  |   rating: Int
10 |   product: Product @provides(fields: "price")
   |                      ^
11 | }
```
it will now show:
```
reviews:10:20
9  |   rating: Int
10 |   product: Product @provides(fields: "price")
   |                      ^
11 | }
```

But this also make slighly more unlikely that this readiness tool
behaviour diverges from the fed2 actual one.