fix: downgrade composition to 2.0.0
Studio still runs 2.0 so users could get a false positive result if they have subgraphs that trigger issues that we've fixed in patch releases

The specific issue this addresses is when subgraphs don't correctly define federation directives.

This also changes `graph_ref` to `graphref` so that the config files are compatible with rover.